Add time display alignment option to select fixed or centered alignment
Implement handling of unique case for alignment option by use of a case class, with a map for relating the class to a preference key. This arrangement enables simple implementation of a class and its addition to the map to handle future unique cases. The technique is discussed in chapter 3, section 3.2 of `Building Maintainable Software, Java Edition` by Joost Visser.